moved test stuff, for better experience with ghcid
