+++ app/Ampersand/Main.hs
@@ -1,60 +1,5 @@
-{-# LANGUAGE RecordWildCards #-}
 module Main(main) where
-
-import           Ampersand
-import qualified RIO.List as L
-import qualified Data.List.NonEmpty as NEL
-import           System.Environment    (getArgs, getProgName)
--- This datatype contains global information available throughout the lifefime of the application.
+import MainApps
 
 main :: IO ()
-main = do
-  opts@Options{..} <- getOptionsIO
-  let app = App opts stdout
-  runRIO app $ do
-     ampersand
-ampersand :: RIO App ()
-ampersand = do
-    app <- ask
-    let opts'@Options{..} = getOptions app
-    sequence_ . map snd . filter fst $ actionsWithoutScript opts'-- There are commands that do not need a single filename to be speciied
-    case fileName of
-      Just _ -> do -- An Ampersand script is provided that can be processed
-            putStrLn "Processing your model..."
-            gMulti <- createMulti
-            case gMulti of
-              Errors err    -> 
-                 exitWith . NoValidFSpec . L.intersperse  (replicate 30 '=') 
-               . fmap show . NEL.toList $ err
-              Checked multi ws -> do
-                 mapM_  putStrLn . concatMap (lines . show) $ ws
-                 generateAmpersandOutput opts' multi
-                 putStrLn "Finished processing your model"
-                 putStrLn . ("Your script has no errors " ++) $
-                    case ws of
-                      []  -> "and no warnings"
-                      [_] -> ", but one warning was found"
-                      _   -> ", but "++show (length ws)++" warnings were found"
-           
-      Nothing -> -- No Ampersand script is provided 
-         if or (map fst $ actionsWithoutScript opts')
-         then verboseLn "No further actions, because no ampersand script is provided"
-         else liftIO $ do 
-            args     <- getArgs
-            progName <- getProgName
-            exitWith . NoAmpersandScript $
-                 [ "No ampersand script provided. Use --help for usage information"
-                 , "   " <> progName <> (concat $ fmap (" " <>) args) ]
-
- where
-   actionsWithoutScript :: Options -> [(Bool, RIO App ())]
-   actionsWithoutScript opts@Options{..} = 
-      [ ( test                     , putStrLn $ "Executable: " ++ show dirExec )
-      , ( showVersion  || verboseP , putStrLn $ versionText opts)
-      , ( genSampleConfigFile      , liftIO writeConfigFile)
-      , ( showHelp                 , putStrLn $ usageInfo' opts)
-      , ( runAsDaemon              , runDaemon)
-      ]
-   
-   versionText :: Options -> String
-   versionText opts = preVersion opts ++ ampersandVersionStr ++ postVersion opts
+main = ampersand
