+++ src/Ampersand/Test/Parser/QuickChecks.hs
@@ -46,11 +46,12 @@ checkArgs = Args
   }
 
 -- TODO: Improve the messages given here, remove all trace's
-test :: Testable prop => prop -> IO (Bool, String)
-test p = do res <- quickCheckWithResult checkArgs p
-            case res of
-                Success {} -> return (True, "")
-                _          -> return (False, show res )
+test :: Testable prop => prop -> RIO env (Bool, String)
+test p = do 
+    res <- liftIO $ quickCheckWithResult checkArgs p
+    case res of
+      Success {} -> return (True, "")
+      _          -> return (False, show res )
 
-parserQuickChecks :: IO (Bool,String)
+parserQuickChecks :: RIO env (Bool,String)
 parserQuickChecks = test prop_pretty
